+++ routes/managementFeeRouter.js
@@ -1,12 +1,25 @@
 const express = require("express");
 const router = express.Router();
 const pool = require("../DB/dbPool");
+const checkServiceKeyResult = require("../modules/authentication");
 
 // 관리비 조회
 router.get("/getMngFeeList", async (req, res, next) => {
-  let { dongCode = "", hoCode = "", hAreaType = "" } = req.query;
-  console.log(dongCode, hoCode, hAreaType);
-
+  let {
+    serviceKey = "111111111", // 서비스 인증키
+    numOfRows = 10, //           페이지 당 결과수
+    pageNo = 1, //               페이지 번호
+    dongCode = "",
+    hoCode = "",
+    hAreaType = "",
+  } = req.query;
+  console.log(serviceKey, numOfRows, pageNo, dongCode, hoCode, hAreaType);
+  if ((await checkServiceKeyResult(serviceKey)) == false) {
+    return res.json({
+      resultCode: "30",
+      resultMsg: "등록되지 않은 서비스키 입니다.",
+    });
+  }
   try {
     let defaultCondition = `LIKE '%'`;
     let defaultDongCondition = defaultCondition;
@@ -30,21 +43,38 @@ router.get("/getMngFeeList", async (req, res, next) => {
       defaultAreaCondition = "";
     }
 
+    let sRow = (pageNo - 1) * numOfRows;
+    let size = numOfRows * 1;
+
     const sql = `SELECT ROW_NUMBER() OVER(ORDER BY insert_date) AS No, a.dong_code AS dongCode, a.ho_code AS hoCode, b.h_area_type AS hAreaType, 
                         CONCAT(mng_year, "-",mng_month) AS payMonth, a.total_mng AS totalMng
                  FROM t_management_fee a
                  INNER JOIN t_dongho b
                  WHERE (a.dong_code = b.dong_code AND a.ho_code = b.ho_code)
                        AND (a.dong_code ${defaultDongCondition} ${dongCondition} AND a.ho_code  ${defaultHoCondition} ${hoCondition})
-                       AND b.h_area_type ${defaultAreaCondition} ${areaCondition};`;
+                       AND b.h_area_type ${defaultAreaCondition} ${areaCondition}
+                       LIMIT ?,?`;
     console.log("sql: " + sql);
-    const data = await pool.query(sql);
+    const data = await pool.query(sql, [Number(sRow), Number(size)]);
+
+    const sql2 = `SELECT count(a.insert_date) as cnt
+                  FROM t_management_fee a
+                  INNER JOIN t_dongho b
+                  WHERE (a.dong_code = b.dong_code AND a.ho_code = b.ho_code)
+                  AND (a.dong_code ${defaultDongCondition} ${dongCondition} AND a.ho_code  ${defaultHoCondition} ${hoCondition})
+                  AND b.h_area_type ${defaultAreaCondition} ${areaCondition};`;
+    const data2 = await pool.query(sql2);
+    let resultCnt = data2[0];
+
     let resultList = data[0];
     let jsonResult = {
       resultCode: "00",
       resultMsg: "NORMAL_SERVICE",
+      numOfRows,
+      pageNo,
+      totalCount: resultCnt[0].cnt + "",
       data: {
-        resultList,
+        items: resultList,
       },
     };
     return res.json(jsonResult);
@@ -58,9 +88,14 @@ router.get("/getMngFeeList", async (req, res, next) => {
 
 // 관리비 삭제
 router.delete("/deleteMngFeeList", async (req, res, next) => {
-  let { dongCode = "", hoCode = "" } = req.body;
-  console.log(dongCode.hoCode);
-
+  let { serviceKey = "", dongCode = "", hoCode = "" } = req.body;
+  console.log(serviceKey, dongCode.hoCode);
+  if ((await checkServiceKeyResult(serviceKey)) == false) {
+    return res.json({
+      resultCode: "30",
+      resultMsg: "등록되지 않은 서비스키 입니다.",
+    });
+  }
   try {
     const sql = `DELETE FROM t_management_fee WHERE dong_code = ? AND ho_code = ?`;
     console.log("sql: " + sql);
